Buildnum и исправление ошибок

+++ src/edit_history.cpp
@@ -10,7 +10,7 @@
 
 
 
-CPropertyChangeAction::CPropertyChangeAction(int serialNumber, VariantValue oldValue, VariantValue newValue)
+CPropertyChangeAction::CPropertyChangeAction(size_t serialNumber, VariantValue oldValue, VariantValue newValue)
 {
 	m_SerialNumber = serialNumber;
 	m_OldValue = oldValue;
